ADD: new tests for debruijn graph chain compaction. FIX: right side end should use complement strand kmers. ENH: changed format of compacted chain node (negative number means opposite strand). ADD: debruijn graph with edge frequencies.  ADD:compacted debruijn graph file output

+++ src/debruijn/debruijn_graph_map.hpp
@@ -187,7 +187,7 @@ namespace bliss {
 
     template<typename Kmer >
     using count_hash_compact_debruijn_graph_map = ::bliss::debruijn::graph::compact_debruijn_graph_map<Kmer,
-    		::bliss::debruijn::graph::compact_multi_biedge<typename Kmer::KmerAlphabet, uint16_t>,
+    		::bliss::debruijn::graph::compact_multi_biedge<typename Kmer::KmerAlphabet, uint32_t>,
         ::bliss::debruijn::CanonicalDeBruijnHashMapParams>;
 
 	  } // namespace graph
